+++ x/node/genesis.go
@@ -12,6 +12,10 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	for _, elem := range genState.NodeList {
 		k.SetNode(ctx, elem)
 	}
+	// Set all the pledgeDebt
+	for _, elem := range genState.PledgeDebtList {
+		k.SetPledgeDebt(ctx, elem)
+	}
 	for _, elem := range genState.PledgeList {
 		k.SetPledge(ctx, elem)
 	}
@@ -32,6 +36,7 @@ func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 		genesis.Pool = &pool
 	}
 	genesis.PledgeList = k.GetAllPledge(ctx)
+	genesis.PledgeDebtList = k.GetAllPledgeDebt(ctx)
 	// this line is used by starport scaffolding # genesis/module/export
 
 	return genesis